# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CFDI Verification API",
    "description": "API para verificar la validez de Comprobantes Fiscales Digitales por Internet (CFDI)",
    "version": "1.0.0"
  },
  "paths": {
    "/verify-cfdi": {
      "post": {
        "tags": [
          "CFDI"
        ],
        "summary": "Verify Cfdi",
        "description": "Verifica la validez de un CFDI con el SAT\n\nEsta API consulta el servicio oficial del SAT para verificar el estatus de un CFDI.\nRequiere autenticaci\u00f3n mediante Bearer token.\n\nReturns:\n    CFDIResponse: Informaci\u00f3n sobre la validez del CFDI",
        "operationId": "verify_cfdi_verify_cfdi_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CFDIRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFDIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/verify-cfdi-batch": {
      "post": {
        "tags": [
          "CFDI"
        ],
        "summary": "Verify Cfdi Batch",
        "description": "Verifica la validez de m\u00faltiples CFDIs con el SAT en una sola petici\u00f3n\n\nEsta API consulta el servicio oficial del SAT para verificar el estatus de m\u00faltiples CFDIs.\nCada CFDI se procesa de forma independiente y los resultados se devuelven en un \u00fanico response.\nRequiere autenticaci\u00f3n mediante Bearer token.\n\nReturns:\n    BatchCFDIResponse: Informaci\u00f3n sobre la validez de todos los CFDIs solicitados",
        "operationId": "verify_cfdi_batch_verify_cfdi_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCFDIRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchCFDIResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/admin/tokens": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create Api Token",
        "description": "Create a new API token\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "create_api_token_admin_tokens_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Api Tokens",
        "description": "List all API tokens\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "list_api_tokens_admin_tokens_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tokens/{token_id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Api Token",
        "description": "Get a specific API token by ID\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "get_api_token_admin_tokens__token_id__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Api Token",
        "description": "Update an API token\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "update_api_token_admin_tokens__token_id__put",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete Api Token",
        "description": "Delete an API token\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "delete_api_token_admin_tokens__token_id__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/tokens/{token_id}/regenerate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Regenerate Api Token",
        "description": "Regenerate an API token\n\nCreates a new token value for the existing token ID.\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "regenerate_api_token_admin_tokens__token_id__regenerate_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Token Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/superadmins": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create New Superadmin",
        "description": "Create a new superadmin\n\nRequires existing superadmin authentication using HTTP Basic auth.",
        "operationId": "create_new_superadmin_admin_superadmins_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuperAdminCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuperAdminResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/admin/superadmins/{username}/password": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update Admin Password",
        "description": "Update a superadmin's password\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "update_admin_password_admin_superadmins__username__password_put",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuperAdminUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/superadmins/{username}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Deactivate Admin Account",
        "description": "Deactivate a superadmin account\n\nRequires superadmin authentication using HTTP Basic auth.",
        "operationId": "deactivate_admin_account_admin_superadmins__username__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CFDIRequest": {
        "properties": {
          "uuid": {
            "type": "string",
            "title": "Uuid",
            "description": "UUID del CFDI",
            "example": "6128396f-c09b-4ec6-8699-43c5f7e3b230"
          },
          "emisor_rfc": {
            "type": "string",
            "title": "Emisor Rfc",
            "description": "RFC del emisor",
            "example": "CDZ050722LA9"
          },
          "receptor_rfc": {
            "type": "string",
            "title": "Receptor Rfc",
            "description": "RFC del receptor",
            "example": "XIN06112344A"
          },
          "total": {
            "type": "string",
            "title": "Total",
            "description": "Monto total del CFDI",
            "example": "12000.00"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "emisor_rfc",
          "receptor_rfc",
          "total"
        ],
        "title": "CFDIRequest"
      },
      "CFDIResponse": {
        "properties": {
          "estado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado",
            "description": "Estado del CFDI"
          },
          "es_cancelable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Es Cancelable",
            "description": "Si el CFDI es cancelable"
          },
          "estatus_cancelacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estatus Cancelacion",
            "description": "Estatus de cancelaci\u00f3n"
          },
          "codigo_estatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codigo Estatus",
            "description": "C\u00f3digo de estatus"
          },
          "validacion_efos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validacion Efos",
            "description": "Validaci\u00f3n EFOS"
          },
          "raw_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Response",
            "description": "Respuesta XML completa"
          }
        },
        "type": "object",
        "title": "CFDIResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}